requirements.txt
mkdocs
mkdocs-material
pyyaml
